requirements.txt
aiogram==2.25.1
black==23.3.0
python-dotenv==1.0.0
